requirements.txt
grpcio==1.2.1
numpy==1.11.1
